UI medicine fix clean code
