+++ packages/doctor_pet/lib/utils/app_assets.dart
@@ -1,3 +1,3 @@
 class AppImages {
-  static const String dcareImage = "assets/img/dcare.png";
+
 }
