+++ packages/doctor_pet/lib/views/doctor/doctor_medicine/doctor_medicine_controller.dart
@@ -2,6 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 import 'package:doctor_pet/core/data/medicine.dart';
 import 'package:doctor_pet/data/data_mock/doctor_medicine.dart';
+import 'package:doctor_pet/views/doctor/doctor_medicine/widgets/add_medicine_dialog.dart';
 
 class DoctorMedicineController extends GetxController {
   Rx<List<Medicine>> medicines = Rx<List<Medicine>>([]);
@@ -15,9 +16,8 @@ class DoctorMedicineController extends GetxController {
   }
 
   Future<void> fetchMedicines() async {
-    await Future.delayed(const Duration(seconds: 1)); // Simulate network delay
-    medicines.value =
-        mockMedicines; // Assuming mockMedicines is your list of medicines
+    await Future.delayed(const Duration(seconds: 1));
+    medicines.value = mockMedicines;
   }
 
   void addSelectedMedicine(Medicine medicine, int quantity, String note) {
@@ -35,108 +35,7 @@ class DoctorMedicineController extends GetxController {
   }
 
   void showAddMedicineDialog() {
-    List<TextEditingController> quantityControllers = [];
-    List<TextEditingController> noteControllers = [];
-    List<Widget> medicineWidgets = [];
-
-    void addMedicineRow() {
-      TextEditingController quantityController = TextEditingController();
-      TextEditingController noteController = TextEditingController();
-      quantityControllers.add(quantityController);
-      noteControllers.add(noteController);
-
-      medicineWidgets.add(
-        Row(
-          children: [
-            Expanded(
-              child: DropdownButtonFormField<Medicine>(
-                decoration: const InputDecoration(labelText: 'Medicine'),
-                value: medicines.value.first,
-                onChanged: (Medicine? value) {},
-                items: medicines.value
-                    .map<DropdownMenuItem<Medicine>>((Medicine medicine) {
-                  return DropdownMenuItem<Medicine>(
-                    value: medicine,
-                    child: Text(medicine.nameMedicine),
-                  );
-                }).toList(),
-              ),
-            ),
-            const SizedBox(width: 10),
-            Expanded(
-              child: TextFormField(
-                controller: quantityController,
-                keyboardType: TextInputType.number,
-                decoration: const InputDecoration(labelText: 'Quantity'),
-              ),
-            ),
-            const SizedBox(width: 10),
-            Expanded(
-              child: TextFormField(
-                controller: noteController,
-                decoration: const InputDecoration(labelText: 'Note'),
-              ),
-            ),
-          ],
-        ),
-      );
-    }
-
-    // Initialize the first set of input fields
-    addMedicineRow();
-
-    Get.dialog(
-      AlertDialog(
-        title: const Text('Add Medicine'),
-        content: StatefulBuilder(
-          builder: (BuildContext context, StateSetter setState) {
-            return SingleChildScrollView(
-              child: Column(
-                mainAxisSize: MainAxisSize.min,
-                children: [
-                  ...medicineWidgets,
-                  const SizedBox(height: 20),
-                  TextButton(
-                    onPressed: () => setState(() => addMedicineRow()),
-                    child: const Text('Add More'),
-                  ),
-                ],
-              ),
-            );
-          },
-        ),
-        actions: [
-          TextButton(
-            onPressed: () => Get.back(),
-            child: const Text('Cancel'),
-          ),
-          TextButton(
-            onPressed: () {
-              for (int i = 0; i < quantityControllers.length; i++) {
-                final Medicine? selectedMedicine =
-                    medicines.value.length > i ? medicines.value[i] : null;
-                if (selectedMedicine != null) {
-                  final int quantity =
-                      int.tryParse(quantityControllers[i].text) ?? 0;
-                  final String note = noteControllers[i].text;
-                  if (quantity > 0) {
-                    addSelectedMedicine(selectedMedicine, quantity, note);
-                  }
-                }
-              }
-              Get.back(); // Close the dialog
-              Get.snackbar(
-                'Success', // Title
-                'The medicine has been successfully saved!', // Message
-                backgroundColor: Colors.green,
-                snackPosition: SnackPosition.BOTTOM,
-                colorText: Colors.white,
-              );
-            },
-            child: const Text('Save'),
-          ),
-        ],
-      ),
-    );
+    Get.dialog(AddMedicineDialog(
+        medicines: medicines, addSelectedMedicine: addSelectedMedicine));
   }
 }
